fix: change svg form icons to jsx components
Found a bug (probably with babel-inline-react-svg) where the fill
url id would be changed to "#a", regardless of what the original id name
was. This was causing a problem with the mega X and mega Y hover on
Charizard page, since he has both forms and hovering one would also
affect the second icon. Manually changing the icons to jsx fixed
this issue and allowed the icons to be implemented.